pyproject.toml
[project]
name = "tortoise-pathway"
version = "0.1.0"
description = "A schema migration tool for Tortoise ORM"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "tortoise-orm>=0.24.2",
]
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: AsyncIO",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://github.com/yourname/tortoise-pathway"
Documentation = "https://github.com/yourname/tortoise-pathway"
Repository = "https://github.com/yourname/tortoise-pathway.git"
Issues = "https://github.com/yourname/tortoise-pathway/issues"

[project.scripts]
tortoise-pathway = "tortoise_pathway.cli:main"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]
dev = [
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tortoise_pathway"]

[tool.black]
line-length = 88
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
